+++ include/SMS/MapObj/MapObjInit.hxx
@@ -60,7 +60,7 @@ struct hit_data {
 struct obj_hit_info {
     u32 _00;
     u32 mType;
-    u32 _08;
+    f32 _08;
     hit_data *mHitData;
 };
